+++ iti-react/Inputs/AddressInput/index.ts
@@ -1,1 +1,2 @@
 ﻿export * from './AddressInput'
+export * from './PostalCodeValidator'